+++ storybook/index.d.ts
@@ -1,1 +1,2 @@
 export declare const pxblueTheme: any;
+export declare const pxblueDarkTheme: any;
